requirements.txt
Django==1.4.3
distribute==0.6.31
dj-database-url==0.2.1
django-haystack==1.2.7
django-tastypie==0.9.12
django-uuidfield==0.4.0
gunicorn==0.17.2
mimeparse==0.1.3
psycopg2==2.4.6
python-dateutil==1.5
wsgiref==0.1.2
